Create anonyfiles_core package and move core functionality
