+++ anonyfiles_core/anonymizer/excel_processor.py
