+++ anonyfiles_core/anonymizer/pdf_processor.py
